+++ src/main/java/akressiopertti/service/FoodStuffService.java
@@ -28,6 +28,9 @@ public class FoodStuffService {
 
     public FoodStuff remove(Long id) {
         FoodStuff foodStuff = foodStuffRepository.findOne(id);
+        if(foodStuff == null) {
+            throw new NullPointerException("Cannot remove object with id " + id.toString());
+        }
         foodStuffRepository.delete(foodStuff);
         return foodStuff;
     }
